#841 added EKMService
